Removed old generic host worker service and added new one
