+++ DependencyInjectionDemos.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28729.10
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenericHostWorkerApp", "GenericHostWorkerApp\GenericHostWorkerApp.csproj", "{38D74F52-9898-449D-A790-DAA59622244C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleContainer.Tests", "SampleContainer.Tests\SampleContainer.Tests.csproj", "{055A41F5-8567-448E-9803-ACB1CD3A0A2E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleContainer", "SampleContainer\SampleContainer.csproj", "{462F2CD0-6123-450E-A9D6-2D439C5403F5}"
@@ -13,16 +11,14 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RetroGamingWebApp", "RetroG
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LeaderboardWebApi", "LeaderboardWebApi\LeaderboardWebApi.csproj", "{6736AB6A-BED7-42E2-94B8-9DF8DDAF0AD9}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HighScoreWorkerService", "HighScoreWorkerService\HighScoreWorkerService.csproj", "{49B96BE0-0B9E-4B65-80D7-3C740E7C7B49}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{38D74F52-9898-449D-A790-DAA59622244C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{38D74F52-9898-449D-A790-DAA59622244C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{38D74F52-9898-449D-A790-DAA59622244C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{38D74F52-9898-449D-A790-DAA59622244C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{055A41F5-8567-448E-9803-ACB1CD3A0A2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{055A41F5-8567-448E-9803-ACB1CD3A0A2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{055A41F5-8567-448E-9803-ACB1CD3A0A2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -39,6 +35,10 @@ Global
 		{6736AB6A-BED7-42E2-94B8-9DF8DDAF0AD9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6736AB6A-BED7-42E2-94B8-9DF8DDAF0AD9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6736AB6A-BED7-42E2-94B8-9DF8DDAF0AD9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{49B96BE0-0B9E-4B65-80D7-3C740E7C7B49}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{49B96BE0-0B9E-4B65-80D7-3C740E7C7B49}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{49B96BE0-0B9E-4B65-80D7-3C740E7C7B49}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{49B96BE0-0B9E-4B65-80D7-3C740E7C7B49}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
