--- GenericHostWorkerApp/GenericHostWorkerApp.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="LightInject.Microsoft.DependencyInjection" Version="2.3.0" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.0.0-preview5.19227.9" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SampleContainer\SampleContainer.csproj" />
-  </ItemGroup>
-</Project>
