--- GenericHostWorkerApp/IThingService.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace GenericHostWorkerApp
-{
-    public interface IThingService
-    {
-        IEnumerable<string> ProduceValues();
-    }
-}
