+++ HighScoreWorkerService/Properties/launchSettings.json
@@ -1,10 +1,13 @@
-﻿{
+{
   "profiles": {
-    "GenericHostWorkerApp": {
+    "HighScoreWorkerService": {
       "commandName": "Project",
       "environmentVariables": {
         "DOTNET_ENVIRONMENT": "Development"
       }
+    },
+    "Docker": {
+      "commandName": "Docker"
     }
   }
 }
